requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.2.0
certifi==2023.7.22
charset-normalizer==2.0.12
colorama==0.4.5
context.api==0.0
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.26
decorator==5.1.1
dill==0.3.5.1
dm-tree==0.1.7
etils==0.6.0
fake-useragent==0.1.14
flatbuffers==2.0
fonttools==4.33.3
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
grpcio==1.53.0
h5py==3.7.0
idna==3.3
imageio==2.28.1
importlib-metadata==4.11.4
importlib-resources==5.8.0
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy_loader==0.2
libclang==14.0.1
loguru==0.6.0
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.3
matplotlib==3.5.2
networkx==3.1
numpy==1.22.0
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.3
parse==1.19.0
Pillow==10.0.1
promise==2.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pydot==1.4.2
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.20.0
scikit-learn==1.1.1
scipy==1.10.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
split-folders==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-addons==0.17.0
tensorflow-datasets==4.6.0
tensorflow-determinism==0.3.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-metadata==1.9.0
tensorflow-model-optimization==0.7.2
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2023.4.12
toml==0.10.2
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.2.0
urllib3==1.26.17
w3lib==2.0.1
websockets==10.4
Werkzeug==2.2.3
win32-setctime==1.1.0
wrapt==1.14.1
zipp==3.8.0
